annotate processors, implement feature switch for cassandra match write
