+++ processors/processMultiKillStreaks.js
@@ -1,5 +1,8 @@
-// associate kill streaks with multi kills and team fights
-module.exports = function processMultiKillStreaks(entries, hero_to_slot, parsed_data)
+/**
+ * Associates kill streaks with multi kills and team fights
+ * THIS PROCESSOR IS CURRENTLY DEPRECATED AND NOT IN PRODUCTION USE (originally implemented by @nickhh)
+ **/
+function processMultiKillStreaks(entries, hero_to_slot, parsed_data)
 {
     var print_multi_kill_streak_debugging = false;
     // bookkeeping about each player
@@ -146,3 +149,4 @@ module.exports = function processMultiKillStreaks(entries, hero_to_slot, parsed_
         }
     }
 }
+module.exports = processMultiKillStreaks;
