+++ processors/processUploadProps.js
@@ -1,4 +1,8 @@
-module.exports = function processUploadProps(entries, meta, populate)
+/**
+ * A processor to extract basic match stats from the replay file.
+ * This is used for uploaded match parses since we can't get basic match data from the API.
+ **/
+function processUploadProps(entries, meta)
 {
     var container = {
         player_map:
@@ -38,4 +42,5 @@ module.exports = function processUploadProps(entries, meta, populate)
         }
     }
     return container;
-};
+}
+module.exports = processUploadProps;
