+++ scanner.js
@@ -6,6 +6,7 @@ var config = require('./config');
 var constants = require('./constants');
 var buildSets = require('./store/buildSets');
 var db = require('./store/db');
+var cassandra = config.ENABLE_CASSANDRA_MATCH_STORE_WRITE ? require('./store/cassandra') : undefined;
 var redis = require('./store/redis');
 var queue = require('./store/queue');
 var queries = require('./store/queries');
@@ -173,6 +174,7 @@ function scanApi(seq_num)
                         {
                             type: "api",
                             origin: "scanner",
+                            cassandra: cassandra,
                         }, close);
                     }
                     else
