+++ store/playerCache.js
@@ -139,9 +139,9 @@ function updateCache(match, cb)
 
 function validateCache(db, redis, account_id, cache, cb)
 {
-    if (!cache || !enabled)
+    if (!cache || !enabled || config.ENABLE_CASSANDRA_MATCH_STORE_READ)
     {
-        console.log('player cache disabled');
+        console.log('player cache validation skipped due to not needed');
         return cb();
     }
     //set key in redis to mark cache audited, don't do it again until timeout
